refactor(sidecar): split MCP listener modules
